requirements.txt
amqp==5.1.0
asgiref==3.5.0
billiard==3.6.4.0
black==22.3.0
celery==5.2.3
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.12
django-celery-beat==2.2.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-nested-routers==0.93.4
flake8==4.0.1
kombu==5.2.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
prompt-toolkit==3.0.28
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.3.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
vine==5.0.0
wcwidth==0.2.5
